requirements.txt
streamlit
torch
torchvision
openai-whisper
pillow
numpy
